refactor: introduce interface for builder
