+++ src/utils/SQLBuilder/utils/builder.ts
@@ -3,9 +3,9 @@ import { PreparedValue } from '../core/PreparedValue';
 import { GroupExpression } from '../GroupExpression';
 import { LimitClause } from '../LimitClause';
 import { SetExpression } from '../SetExpression';
+import { RawQueryParameter } from '../types';
 import { WhereClause } from '../WhereClause';
 import { QueryConstructor } from './QueryConstructor';
-import { RawQueryParameter } from '..';
 
 /**
  * SQL query builder
